+++ qupath-core/src/main/java/qupath/lib/objects/hierarchy/PathObjectHierarchy.java
@@ -26,6 +26,7 @@ package qupath.lib.objects.hierarchy;
 import java.io.Serial;
 import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
@@ -36,11 +37,13 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import qupath.lib.analysis.DelaunayTools;
 import qupath.lib.common.LogTools;
 import qupath.lib.objects.DefaultPathObjectComparator;
 import qupath.lib.objects.PathAnnotationObject;
@@ -55,6 +58,7 @@ import qupath.lib.objects.hierarchy.events.PathObjectHierarchyEvent;
 import qupath.lib.objects.hierarchy.events.PathObjectHierarchyListener;
 import qupath.lib.objects.hierarchy.events.PathObjectSelectionModel;
 import qupath.lib.objects.hierarchy.events.PathObjectHierarchyEvent.HierarchyEventType;
+import qupath.lib.regions.ImagePlane;
 import qupath.lib.regions.ImageRegion;
 import qupath.lib.roi.interfaces.ROI;
 
@@ -88,6 +92,9 @@ public final class PathObjectHierarchy implements Serializable {
 
 	// Cache enabling faster access of objects according to location
 	private final transient PathObjectTileCache tileCache = new PathObjectTileCache(this);
+
+	// A map to store subdivisions, useful for finding neighbors
+	private transient SubdivisionManager subdivisionManager = new SubdivisionManager();
 
 	/**
 	 * Default constructor, creates an empty hierarchy.
@@ -278,7 +285,7 @@ public final class PathObjectHierarchy implements Serializable {
 			logger.warn("TMA core objects cannot be inserted - use resolveHierarchy() instead");
 			return false;
 		}
-		
+
 		// Get all the annotations that might be a parent of this object
 		var region = ImageRegion.createInstance(pathObject.getROI());
 		Collection<PathObject> tempSet = new HashSet<>();
@@ -442,25 +449,6 @@ public final class PathObjectHierarchy implements Serializable {
 			addPathObjectImpl(pathObject, false);
 		}
 		fireHierarchyChangedEvent(this);
-		
-		// This previously could result in child objects being deleted even if keepChildren was 
-		// true, depending upon the order in which objects were removed.
-//		// Loop through and remove objects
-//		for (Entry<PathObject, List<PathObject>> entry : map.entrySet()) {
-//			PathObject parent = entry.getKey();
-//			List<PathObject> children = entry.getValue();
-//			parent.removePathObjects(children);
-//			if (keepChildren) {
-//				for (PathObject child : children) {
-//					if (child.hasChildren()) {
-//						List<PathObject> newChildList = new ArrayList<>(child.getChildObjects());
-//						newChildList.removeAll(pathObjects);
-//						parent.addPathObjects(newChildList);
-//					}
-//				}
-//			}
-//		}
-//		fireHierarchyChangedEvent(this);
 	}
 	
 	
@@ -561,9 +549,9 @@ public final class PathObjectHierarchy implements Serializable {
 			changes = addPathObjectToList(getRootObject(), pathObject, false) || changes;
 			counter++;
 		}
-		if (changes)
+		if (changes) {
 			fireHierarchyChangedEvent(getRootObject());
-//			fireChangeEvent(getRootObject());
+		}
 		return changes;
 	}
 
@@ -1139,15 +1127,163 @@ public final class PathObjectHierarchy implements Serializable {
 	
 	synchronized void fireEvent(PathObjectHierarchyEvent event) {
 		synchronized(listeners) {
+			if (!event.isChanging()) {
+				if (event.isStructureChangeEvent()) {
+					var changed = event.getChangedObjects();
+					var classes = changed.stream().map(PathObject::getClass).distinct().toList();
+					if (classes.isEmpty() || classes.contains(PathRootObject.class))
+						resetNeighbors();
+					else {
+						for (var cls : classes) {
+							resetNeighborsForClass(cls);
+						}
+					}
+				}
+			}
+
 			for (PathObjectHierarchyListener listener : listeners)
 				listener.hierarchyChanged(event);
 		}
 	}
-	
+
+	private synchronized void resetNeighborsForClass(Class<? extends PathObject> cls) {
+		subdivisionManager.clear();
+	}
+
+	private synchronized void resetNeighbors() {
+		subdivisionManager.clear();
+	}
+
+	/**
+	 * Find all neighbors of a PathObject, having the same class as the object (e.g. detection, cell, annotation).
+	 * This is based on centroids and Delaunay triangulation.
+	 * It also assumes 'square' pixels, and searches for neighbors only on the same 2D plane (z and t).
+	 * @param pathObject
+	 * <p>
+	 * This is an <i>experimental method</i> added in v0.6.0, subject to change.
+	 * @return
+	 * @since v0.6.0
+	 */
+	public synchronized List<PathObject> findAllNeighbors(PathObject pathObject) {
+		var subdivision = getSubdivision(pathObject);
+		return subdivision == null ? Collections.emptyList() : subdivision.getNeighbors(pathObject);
+	}
+
+	/**
+	 * Find the nearest neighbor of a PathObject, having the same class as the object (e.g. detection, cell, annotation).
+	 * This is based on centroids and Delaunay triangulation.
+	 * It also assumes 'square' pixels, and searches for neighbors only on the same 2D plane (z and t).
+	 * <p>
+	 * This is an <i>experimental method</i> added in v0.6.0, subject to change.
+	 * @param pathObject
+	 * @return
+	 * @since v0.6.0
+	 */
+	public synchronized PathObject findNearestNeighbor(PathObject pathObject) {
+		var subdivision = getSubdivision(pathObject);
+		return subdivision == null ? null : subdivision.getNearestNeighbor(pathObject);
+	}
+
+	/**
+	 * Get the subdivision containing a specific PathObject.
+	 * This is based on centroids and Delaunay triangulation in 2D.
+	 * It supports #findAllNeighbors(PathObject) and #findNearestNeighbor(PathObject); obtaining the subdivision
+	 * enables a wider range of spatial queries.
+	 * <p>
+	 * This is an <i>experimental method</i> added in v0.6.0, subject to change.
+	 * @param pathObject
+	 * @return
+	 * @since v0.6.0
+	 */
+	public synchronized DelaunayTools.Subdivision getSubdivision(PathObject pathObject) {
+		return subdivisionManager.getSubdivision(pathObject);
+	}
+
+	/**
+	 * Get a subdivision containing detections.
+	 * This does <i>not<</i> include sub-classes such as 'cell' or 'tile'.
+	 * <p>
+	 * This is an <i>experimental method</i> added in v0.6.0, subject to change.
+	 * @param plane
+	 * @return
+	 * @since v0.6.0
+	 */
+	public synchronized DelaunayTools.Subdivision getDetectionSubdivision(ImagePlane plane) {
+		return subdivisionManager.getSubdivision(PathDetectionObject.class, plane);
+	}
+
+	/**
+	 * Get a subdivision containing cell objects.
+	 * <p>
+	 * This is an <i>experimental method</i> added in v0.6.0, subject to change.
+	 * @param plane
+	 * @return
+	 * @since v0.6.0
+	 */
+	public synchronized DelaunayTools.Subdivision getCellSubdivision(ImagePlane plane) {
+		return subdivisionManager.getSubdivision(PathCellObject.class, plane);
+	}
+
+	/**
+	 * Get a subdivision containing annotation objects.
+	 * <p>
+	 * This is an <i>experimental method</i> added in v0.6.0, subject to change.
+	 * @param plane
+	 * @return
+	 * @since v0.6.0
+	 */
+	public synchronized DelaunayTools.Subdivision getAnnotationSubdivision(ImagePlane plane) {
+		return subdivisionManager.getSubdivision(PathAnnotationObject.class, plane);
+	}
+
+
+	private DelaunayTools.Subdivision computeSubdivision(Class<? extends PathObject> cls) {
+		var pathObjects = tileCache.getObjectsForRegion(cls,
+				null, null, false);
+		return DelaunayTools.createFromCentroids(pathObjects, true);
+	}
+
 	
 	@Override
 	public String toString() {
 		return "Hierarchy: " + nObjects() + " objects";
+	}
+
+	private class SubdivisionManager {
+
+		private static final DelaunayTools.Subdivision EMPTY = DelaunayTools.createFromCentroids(Collections.emptyList(), true);
+
+		private static final Map<Class<? extends PathObject>,
+				Map<ImagePlane, DelaunayTools.Subdivision>> subdivisionMap = new ConcurrentHashMap<>();
+
+		synchronized DelaunayTools.Subdivision getSubdivision(PathObject pathObject) {
+			if (pathObject == null || !pathObject.hasROI()) {
+				return EMPTY;
+			}
+			return getSubdivision(pathObject.getClass(), pathObject.getROI().getImagePlane());
+		}
+
+		synchronized DelaunayTools.Subdivision getSubdivision(Class<? extends PathObject> cls, ImagePlane plane) {
+			var map = subdivisionMap.computeIfAbsent(cls, k -> new ConcurrentHashMap<>());
+			return map.computeIfAbsent(plane, k -> computeSubdivision(cls, plane));
+		}
+
+		private DelaunayTools.Subdivision computeSubdivision(Class<? extends PathObject> cls, ImagePlane plane) {
+			var pathObjects = tileCache.getObjectsForRegion(cls,
+					ImageRegion.createInstance(-Integer.MAX_VALUE/2, -Integer.MAX_VALUE/2,
+							Integer.MAX_VALUE, Integer.MAX_VALUE, plane.getZ(), plane.getT()),
+					null, false);
+			return DelaunayTools.createFromCentroids(pathObjects, true);
+		}
+
+		private synchronized void clear() {
+			subdivisionMap.clear();
+		}
+
+		private synchronized void clearClass(Class<? extends PathObject> cls) {
+			subdivisionMap.getOrDefault(cls, Collections.emptyMap()).clear();
+		}
+
 	}
 	
 }